Initialize Corpus Query Processor (CQP).

CQP needs to know where to look for CWB indexed corpora. To initialize CQP, call cqp_initialize. To reset the registry, use the function cqp_reset_registry. To get the registry used by CQP, use cqp_get_registry. To get the initialization status, use cqp_is_initialized

cqp_initialize(registry = Sys.getenv("CORPUS_REGISTRY"))

cqp_is_initialized()

cqp_verbosity(silent, verbose)

cqp_get_registry()

cqp_reset_registry(registry = Sys.getenv("CORPUS_REGISTRY"))

cqp_load_corpus(corpus, registry)

Arguments

registry

the registry directory

silent

A single logical value, whether to be silent and suppress CQP messages (TRUE), or not (FALSE).

verbose

A single logical value, whether to show verbose parser output (TRUE) or not (FALSE).

corpus

ID of a CWB corpus (length-one character).

Details

cqp_load_corpus will return a logical value - TRUE if corpus has been loaded successfully, FALSE if not.

Author

Andreas Blaette, Bernard Desgraupes, Sylvain Loiseau

Examples

cqp_is_initialized() # check initialization status
#> [1] TRUE
if (!cqp_is_initialized()) cqp_initialize()
cqp_is_initialized() # check initialization status (TRUE now?)
#> [1] TRUE
cqp_get_registry() # get registry dir used by CQP
#> /Users/andreasblatte/Data/cwb/registry
cqp_list_corpora() # get list of corpora
#>   [1] "UNGA"                      "REUTERS"                  
#>   [3] "RHLA"                      "HST"                      
#>   [5] "GIAN"                      "GERMAPARLSAMPLE"          
#>   [7] "MBGA"                      "PARTYPAGES"               
#>   [9] "MATK"                      "RZTG"                     
#>  [11] "SAAR"                      "NEUW"                     
#>  [13] "BEZE"                      "NAZT"                     
#>  [15] "SOMP"                      "TA"                       
#>  [17] "SWT"                       "WOAZ"                     
#>  [19] "PAZ"                       "VBW"                      
#>  [21] "NGVZ"                      "MBVS"                     
#>  [23] "BKR"                       "BKU"                      
#>  [25] "TZH"                       "DAW"                      
#>  [27] "FNP"                       "BMP"                      
#>  [29] "SWAZ"                      "MIGPARL"                  
#>  [31] "WEST"                      "GEZE"                     
#>  [33] "LBN"                       "BADZ"                     
#>  [35] "GTB"                       "LR"                       
#>  [37] "OHEZ"                      "NNN"                      
#>  [39] "BGM"                       "HAZ"                      
#>  [41] "DNN"                       "BR"                       
#>  [43] "WAZ"                       "GERMAPARL_PARLACLARIN_III"
#>  [45] "LAZE"                      "GERMAPARL"                
#>  [47] "NPHA"                      "MPW"                      
#>  [49] "WBN"                       "KIRZ"                     
#>  [51] "BZ"                        "WBZ"                      
#>  [53] "OSZ"                       "WEWA"                     
#>  [55] "KN"                        "RELO"                     
#>  [57] "MUVB"                      "HA"                       
#>  [59] "KR"                        "WKR"                      
#>  [61] "NOW"                       "TLZ"                      
#>  [63] "NRZ"                       "MASP"                     
#>  [65] "BUER"                      "MUAZ"                     
#>  [67] "KSTA"                      "ALLZ"                     
#>  [69] "OAZ"                       "OEFA"                     
#>  [71] "MUME"                      "SZO"                      
#>  [73] "AAZ"                       "HIAZ"                     
#>  [75] "STAG"                      "FEPR"                     
#>  [77] "NADIRAFAZ"                 "TV"                       
#>  [79] "DECH"                      "WK"                       
#>  [81] "HOFZ"                      "BVH"                      
#>  [83] "MIB"                       "GERMAPARL2"               
#>  [85] "BEZG"                      "ALBB"                     
#>  [87] "SEGZ"                      "EXPR"                     
#>  [89] "LVZ"                       "WRU"                      
#>  [91] "GEA"                       "COBU"                     
#>  [93] "UNGA"                      "SWP"                      
#>  [95] "RAMOMEDIA_FAZ"             "WESP"                     
#>  [97] "VZS"                       "HATA"                     
#>  [99] "NADIRATAZ"                 "DAZ"                      
#> [101] "NNP"                       "NKU"                      
#> [103] "MAER"                      "NKR"                      
#> [105] "HNA"                       "IDZL"                     
#> [107] "SAZE"                      "BNA"                      
#> [109] "SGZ"                       "NZ"                       
#> [111] "EZE"                       "MUZE"                     
#> [113] "LIN"                       "NN"                       
#> [115] "NOZ"                       "HOZG"                     
#> [117] "OVZ"                       "NADIRASZ"                 
#> [119] "RHUN"                      "RAMOMEDIA_SZ"             
#> [121] "MZ"                        "HK"                       
#> [123] "WORZ"                      "MAZ"                      
#> [125] "SVZ"                       "PRIG"                     
#> [127] "GAZ"                       "NWZ"                      
#> [129] "KRAN"                      "RLNR"                     
#> [131] "DIKI"                      "GERMAPARL_BETA"           
#> [133] "FRT"                       "HAKU"                     
#> [135] "USAN"                      "MOPO"                     
#> [137] "LAAN"                      "BRZE"                     
#> [139] "SK"