Get Matrix with Regions for Strucs.

The return value is an integer matrix with the left and right corpus positions of the strucs in columns one and two, respectively.

get_region_matrix(corpus, s_attribute, strucs,
  registry = Sys.getenv("CORPUS_REGISTRY"))

Arguments

corpus
a CWB corpus

s_attribute
a structural attribute

strucs
strucs

registry
the registry directory

Value

A matrix with integer values indicating left and right corpus positions (columns 1 and 2, respectively).

Examples

registry <- if (!check_pkg_registry_files()) use_tmp_registry() else get_pkg_registry()
y <- get_region_matrix(
  corpus = "REUTERS", s_attribute = "id",
  strucs = 0L:5L, registry = registry
  )