CWB Tools for Creating Corpora

Wrappers for the CWB tools (cwb-makeall, cwb-huffcode, cwb-compress-rdx). Unlike the 'original' command line tools, these wrappers will always perform a specific indexing/compression step on one positional attribute, and produce all components.

cwb_makeall(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"))

cwb_huffcode(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"))

cwb_compress_rdx(corpus, p_attribute, registry = Sys.getenv("CORPUS_REGISTRY"))

cwb_encode(
  corpus,
  registry = Sys.getenv("CORPUS_REGISTRY"),
  data_dir,
  vrt_dir,
  p_attributes = c("word", "pos", "lemma"),
  s_attributes
)

Arguments

corpus

name of a CWB corpus (upper case)

p_attribute

name p-attribute

registry

path to the registry directory, defaults to the value of the environment variable CORPUS_REGISTRY

data_dir

The data directory where cwb_encode will put the binary files of the indexed corpus.

vrt_dir

Directory with input corpus files (verticalised format / file ending *.vrt).

p_attributes

Positional attributes (p-attributes) to be declared.

s_attributes

A list of named character vectors to declare structural attributes that shall be encoded. The names of the list are the XML elements present in the corpus. Character vectors making up the list declare the attributes that include the metadata of regions. To declare a structural attribute without annotations, provide a zero-length character vector using character() - see examples.

Examples

# The package includes and 'unfinished' corpus of debates in the UN General 
# Assembly ("UNGA"), i.e. it does not yet include the reverse index, and it is
# not compressed.
#
# The first step in the following example is to copy the raw
# corpus to a temporary place.

registry <- if (!check_pkg_registry_files()) use_tmp_registry() else get_pkg_registry()
home_dir <- system.file(package = "RcppCWB", "extdata", "cwb", "indexed_corpora", "unga")

tmpdir <- normalizePath(tempdir(), winslash = "/")
tmp_regdir <- file.path(tmpdir, "registry_tmp", fsep = "/")
tmp_data_dir <- file.path(tmpdir, "indexed_corpora", fsep = "/")
tmp_unga_dir <- file.path(tmp_data_dir, "unga", fsep = "/")
if (!file.exists(tmp_regdir)) dir.create(tmp_regdir)
if (!file.exists(tmp_data_dir)) dir.create(tmp_data_dir)
if (!file.exists(tmp_unga_dir)){
  dir.create(tmp_unga_dir)
} else {
  file.remove(list.files(tmp_unga_dir, full.names = TRUE))
}
regfile <- readLines(file.path(registry, "unga"))
regfile[grep("^HOME", regfile)] <- sprintf('HOME "%s"', tmp_unga_dir)
writeLines(text = regfile, con = file.path(tmp_regdir, "unga"))
for (x in list.files(home_dir, full.names = TRUE)){
  file.copy(from = x, to = tmp_unga_dir)
}

# perform cwb_makeall (equivalent to cwb-makeall command line utility)
cwb_makeall(corpus = "UNGA", p_attribute = "word", registry = tmp_regdir)
#> === Makeall: processing corpus UNGA ===
#> Registry directory: /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/registry_tmp
#> ATTRIBUTE word
#>  + creating LEXSRT ... OK
#>  - lexicon      OK
#>  + creating FREQS ... OK
#>  - frequencies  OK
#>  - token stream OK
#>  + creating REVCIDX ... OK
#>  + creating REVCORP ... OK
#>  ? validating REVCORP ... OK
#>  - index        OK
#> ========================================
#> [1] 0

# see whether it works
ids_sentence_1 <- cl_cpos2id(
  corpus = "UNGA", p_attribute = "word", registry = tmp_regdir,
  cpos = 0:83
  )
tokens_sentence_1 <- cl_id2str(
  corpus = "UNGA", p_attribute = "word",
  registry = tmp_regdir, id = ids_sentence_1
  )
sentence <- gsub("\\s+([\\.,])", "\\1", paste(tokens_sentence_1, collapse = " "))

# perform cwb_huffcode (equivalent to cwb-makeall command line utility)
if (.Platform$OS.type != "windows"){
  cwb_huffcode(corpus = "UNGA", p_attribute = "word", registry = tmp_regdir)
}
#> COMPRESSING TOKEN STREAM of (null).word
#> Allocated heap with 15902 cells for 7951 items
#> 
#> Minimal code length:   4
#> Maximal code length:  17
#> Compressed code len:    1163628 bits,     145453 (+1) bytes
#> 
#> 
#>  CL  #codes  MinCode   SymIdx
#> ----------------------------------------
#>   4       2       14        0
#>   5       5       23        2
#>   6       1       45        7
#>   7       8       82        8
#>   8      20      144       16
#>   9      38      250       36
#>  10      94      406       74
#>  11     199      613      168
#>  12     351      875      367
#>  13     625     1125      718
#>  14     873     1377     1343
#>  15    1373     1381     2216
#>  16    1162     1600     3589
#>  17    3200        0     4751
#> ----------------------------------------
#>        7951
#> - writing code descriptor block to /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/indexed_corpora/unga/word.hcd
#> - writing compressed item sequence to /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/indexed_corpora/unga/word.huf
#> - writing sync (every 128 tokens) to /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/indexed_corpora/unga/word.huf.syn
#> VALIDATING UNGA.word
#> - reading code descriptor block from /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/indexed_corpora/unga/word.hcd
#> - reading compressed item sequence from /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/indexed_corpora/unga/word.huf
#> - reading sync (mod 128) from /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/indexed_corpora/unga/word.huf.syn
#> !! You can delete the file </private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/indexed_corpora/unga/word.corpus> now.
#> [1] 0
if (.Platform$OS.type != "windows"){
  cwb_compress_rdx(corpus = "UNGA", p_attribute = "word", registry = tmp_regdir)
}
#> COMPRESSING INDEX of UNGA.word
#> - writing compressed index to /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/indexed_corpora/unga/word.crc
#> - writing compressed index offsets to /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/indexed_corpora/unga/word.crx
#> VALIDATING UNGA.word
#> - reading compressed index from /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/indexed_corpora/unga/word.crc
#> - reading compressed index offsets from /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/indexed_corpora/unga/word.crx
#> !! You can delete the file </private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/indexed_corpora/unga/word.corpus.rev> now.
#> !! You can delete the file </private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/indexed_corpora/unga/word.corpus.rdx> now.
#> [1] 0
if (.Platform$OS.type != "windows"){
data_dir <- file.path(tempdir(), "tmp_data_dir")
dir.create(data_dir)

cwb_encode(
  corpus = "BTMIN",
  registry = Sys.getenv("CORPUS_REGISTRY"),
  vrt_dir = system.file(package = "RcppCWB", "extdata", "vrt"),
  data_dir = data_dir,
  p_attributes = c("word", "pos", "lemma"),
  s_attributes = list(
    plenary_protocol = c(
      "lp", "protocol_no", "date", "year", "birthday", "version",
      "url", "filetype"
    ),
    speaker = c(
      "id", "type", "lp", "protocol_no", "date", "year", "ai_no", "ai_id",
      "ai_type", "who", "name", "parliamentary_group", "party", "role"
     ),
    p = character()
  )
)

unlink(data_dir)
unlink(file.path(Sys.getenv("CORPUS_REGISTRY"), "btmin"))
}
#> ATT: -V plenary_protocol:0+lp+protocol_no+date+year+birthday+version+url+filetype
#> ATT: -V plenary_protocol_lp:0
#> ATT: -V plenary_protocol_protocol_no:0
#> ATT: -V plenary_protocol_date:0
#> ATT: -V plenary_protocol_year:0
#> ATT: -V plenary_protocol_birthday:0
#> ATT: -V plenary_protocol_version:0
#> ATT: -V plenary_protocol_url:0
#> ATT: -V plenary_protocol_filetype:0
#> ATT: -V speaker:0+id+type+lp+protocol_no+date+year+ai_no+ai_id+ai_type+who+name+parliamentary_group+party+role
#> ATT: -V speaker_id:0
#> ATT: -V speaker_type:0
#> ATT: -V speaker_lp:0
#> ATT: -V speaker_protocol_no:0
#> ATT: -V speaker_date:0
#> ATT: -V speaker_year:0
#> ATT: -V speaker_ai_no:0
#> ATT: -V speaker_ai_id:0
#> ATT: -V speaker_ai_type:0
#> ATT: -V speaker_who:0
#> ATT: -V speaker_name:0
#> ATT: -V speaker_parliamentary_group:0
#> ATT: -V speaker_party:0
#> ATT: -V speaker_role:0
#> ATT: -S p
#> List of input files:
#>  - /Users/runner/work/_temp/Library/RcppCWB/extdata/vrt/BT_18_001.vrt
#>  - /Users/runner/work/_temp/Library/RcppCWB/extdata/vrt/BT_18_002.vrt
#> Start: Fri Jan 28 19:35:18 2022
#> 
#>         0k tokens processed        0k tokens processed        0k tokens processed        0k tokens processed        0k tokens processed        0k tokens processed        0k tokens processed       14k tokens processed       29k tokens processed       43k tokens processed       43k tokens processed       43k tokens processed                                                  Total size: 50620 tokens (0.0M)
#> Writing registry file /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpoFGpnC/registry_tmp/btmin ...
#> Done: Fri Jan 28 19:35:18 2022
#>